+++ app/configurator/components/ui-helpers.ts
@@ -21,7 +21,7 @@ import {
   TemporalDimension,
   TemporalEntityDimension,
 } from "@/domain/data";
-import { TimeUnit } from "@/graphql/query-hooks";
+import { RelatedDimensionType, TimeUnit } from "@/graphql/query-hooks";
 import { IconName } from "@/icons";
 import { getTimeInterval } from "@/intervals";
 import { getPalette } from "@/palettes";
@@ -107,13 +107,17 @@ export const getTimeIntervalFormattedSelectOptions = ({
 };
 
 export const getErrorMeasure = (
-  { dimensions, measures }: Pick<BaseChartProps, "dimensions" | "measures">,
+  {
+    dimensions,
+    measures,
+    type,
+  }: Pick<BaseChartProps, "dimensions" | "measures"> & {
+    type: RelatedDimensionType;
+  },
   valueIri: string
 ) => {
   return [...dimensions, ...measures].find((m) => {
-    return m.related?.some(
-      (r) => r.type === "StandardError" && r.id === valueIri
-    );
+    return m.related?.some((r) => r.type === type && r.id === valueIri);
   });
 };
 
@@ -122,14 +126,16 @@ export const useErrorMeasure = (
   {
     dimensions,
     measures,
+    type,
   }: {
     dimensions: Dimension[];
     measures: Measure[];
+    type: RelatedDimensionType;
   }
 ) => {
   return useMemo(() => {
-    return getErrorMeasure({ dimensions, measures }, componentId);
-  }, [componentId, dimensions, measures]);
+    return getErrorMeasure({ dimensions, measures, type }, componentId);
+  }, [componentId, dimensions, measures, type]);
 };
 
 export const useErrorVariable = (errorMeasure?: Component) => {
@@ -143,26 +149,38 @@ export const useErrorVariable = (errorMeasure?: Component) => {
 };
 
 export const useErrorRange = (
-  errorMeasure: Component | undefined,
+  upperErrorMeasure: Component | undefined,
+  lowerErrorMeasure: Component | undefined,
   valueGetter: (d: Observation) => number | null
 ) => {
   return useMemo(() => {
-    return errorMeasure
+    return upperErrorMeasure && lowerErrorMeasure
       ? (d: Observation) => {
           const v = valueGetter(d) as number;
-          const errorIri = errorMeasure.id;
-          let error =
-            d[errorIri] !== null ? parseFloat(d[errorIri] as string) : null;
-          if (errorMeasure.unit === "%" && error !== null) {
-            error = (error * v) / 100;
+          const upperId = upperErrorMeasure.id;
+          let upperError =
+            d[upperId] !== null ? parseFloat(d[upperId] as string) : null;
+
+          if (upperErrorMeasure.unit === "%" && upperError !== null) {
+            upperError = (upperError * v) / 100;
           }
-          return (error === null ? [v, v] : [v - error, v + error]) as [
-            number,
-            number,
-          ];
+
+          const lowerId = lowerErrorMeasure.id;
+          let lowerError =
+            d[lowerId] !== null ? parseFloat(d[lowerId] as string) : null;
+
+          if (lowerErrorMeasure.unit === "%" && lowerError !== null) {
+            lowerError = (lowerError * v) / 100;
+          }
+
+          return (
+            upperError === null || lowerError === null
+              ? [v, v]
+              : [v - lowerError, v + upperError]
+          ) as [number, number];
         }
       : null;
-  }, [errorMeasure, valueGetter]);
+  }, [lowerErrorMeasure, upperErrorMeasure, valueGetter]);
 };
 
 export const getIconName = (name: string): IconName => {