+++ app/domain/data.ts
@@ -180,6 +180,18 @@ const ComponentsRenderingConfig: {
     enableMultiFilter: false,
     enableSegment: false,
   },
+  ConfidenceUpperBoundDimension: {
+    enableAnimation: false,
+    enableCustomSort: false,
+    enableMultiFilter: false,
+    enableSegment: false,
+  },
+  ConfidenceLowerBoundDimension: {
+    enableAnimation: false,
+    enableCustomSort: false,
+    enableMultiFilter: false,
+    enableSegment: false,
+  },
 };
 
 export const ANIMATION_ENABLED_COMPONENTS = Object.entries(
@@ -261,7 +273,9 @@ export type Dimension =
   | TemporalOrdinalDimension
   | GeoCoordinatesDimension
   | GeoShapesDimension
-  | StandardErrorDimension;
+  | StandardErrorDimension
+  | ConfidenceUpperBoundDimension
+  | ConfidenceLowerBoundDimension;
 
 export type DimensionType = Dimension["__typename"];
 
@@ -314,6 +328,14 @@ export type GeoShapesDimension = BaseDimension & {
 
 export type StandardErrorDimension = BaseDimension & {
   __typename: "StandardErrorDimension";
+};
+
+export type ConfidenceUpperBoundDimension = BaseDimension & {
+  __typename: "ConfidenceUpperBoundDimension";
+};
+
+export type ConfidenceLowerBoundDimension = BaseDimension & {
+  __typename: "ConfidenceLowerBoundDimension";
 };
 
 export type Measure = NumericalMeasure | OrdinalMeasure;
@@ -610,7 +632,15 @@ export const isTemporalDimensionWithTimeUnit = (
 };
 
 const isStandardErrorResolvedDimension = (dim: ResolvedDimension) => {
-  return dim.data?.related.some((x) => x.type === "StandardError");
+  return dim.data?.related.some((r) => r.type === "StandardError");
+};
+
+const isConfidenceLowerBoundResolvedDimension = (dim: ResolvedDimension) => {
+  return dim.data?.related.some((r) => r.type === "ConfidenceLowerBound");
+};
+
+const isConfidenceUpperBoundResolvedDimension = (dim: ResolvedDimension) => {
+  return dim.data?.related.some((r) => r.type === "ConfidenceUpperBound");
 };
 
 export const isStandardErrorDimension = (
@@ -619,21 +649,27 @@ export const isStandardErrorDimension = (
   return dim.__typename === "StandardErrorDimension";
 };
 
-export const findRelatedErrorDimension = (
-  dimIri: string,
-  dimensions: Component[]
-) => {
-  return dimensions.find((x) =>
-    x.related?.some((r) => r.id === dimIri && r.type === "StandardError")
-  );
+export const isConfidenceUpperBoundDimension = (
+  dim: Component
+): dim is ConfidenceUpperBoundDimension => {
+  return dim.__typename === "ConfidenceUpperBoundDimension";
+};
+
+export const isConfidenceLowerBoundDimension = (
+  dim: Component
+): dim is ConfidenceLowerBoundDimension => {
+  return dim.__typename === "ConfidenceLowerBoundDimension";
 };
 
 export const shouldLoadMinMaxValues = (dim: ResolvedDimension) => {
   const {
     data: { isNumerical, scaleType, dataKind },
   } = dim;
+
   return (
     (isNumerical && scaleType !== "Ordinal" && dataKind !== "Time") ||
-    isStandardErrorResolvedDimension(dim)
+    isStandardErrorResolvedDimension(dim) ||
+    isConfidenceUpperBoundResolvedDimension(dim) ||
+    isConfidenceLowerBoundResolvedDimension(dim)
   );
 };